v1.01 (incomplete)
Refactor of the code generated to provide readability of the code.
Adds name variable (user input) to use as customisation in the output distribution of the budget for the user. 
Uses py .format function for the generation of the output distribution for the organisation of the variables used to calculate the budgetary distribution.